Starting work on FavoriteAdapter. Most complex one by far
