+++ app/src/main/java/fr/neamar/kiss/forwarder/ForwarderManager.java
@@ -14,6 +14,7 @@ public class ForwarderManager extends Forwarder {
     private final WallpaperForwarder wallpaperForwarder;
     private final UITweaksForwarder uiTweaksForwarder;
     private final UXTweaksForwarder uxTweaksForwarder;
+    private final FavoriteForwarder favoriteForwarder;
 
     public ForwarderManager(MainActivity mainActivity, SharedPreferences prefs) {
         super(mainActivity, prefs);
@@ -22,7 +23,7 @@ public class ForwarderManager extends Forwarder {
         this.wallpaperForwarder = new WallpaperForwarder(mainActivity, prefs);
         this.uiTweaksForwarder = new UITweaksForwarder(mainActivity, prefs);
         this.uxTweaksForwarder = new UXTweaksForwarder(mainActivity, prefs);
-
+        this.favoriteForwarder = new FavoriteForwarder(mainActivity, prefs);
     }
 
     @Override
@@ -30,11 +31,13 @@ public class ForwarderManager extends Forwarder {
         super.onCreate();
         widgetForwarder.onCreate();
         uiTweaksForwarder.onCreate();
+        favoriteForwarder.onCreate();
     }
 
     @Override
     public void onResume() {
         uiTweaksForwarder.onResume();
+        favoriteForwarder.onResume();
     }
 
     @Override
@@ -72,7 +75,14 @@ public class ForwarderManager extends Forwarder {
         widgetForwarder.onDataSetChanged();
     }
 
+    @Override
     public void updateRecords(String query) {
         uxTweaksForwarder.updateRecords(query);
     }
+
+    @Override
+    public void allProvidersHaveLoaded() {
+        favoriteForwarder.allProvidersHaveLoaded();
+    }
+
 }
